requirements.txt
# Install these packages with pip
openai==0.11.7
python-dotenv==0.19.2
